address comment on the PR
