+++ services/search.service.ts
@@ -1,6 +1,11 @@
 import { ApiClient } from './client';
-import { SearchSuggestion, transformSearchSuggestion, EntityType } from '@/types/search';
-import { SuggestedAuthor, transformAuthorSuggestions } from '@/types/authorProfile';
+import {
+  SearchSuggestion,
+  transformSearchSuggestion,
+  EntityType,
+  AuthorSuggestion,
+} from '@/types/search';
+import { transformAuthorSuggestions } from '@/types/search';
 
 export class SearchService {
   private static readonly BASE_PATH = '/api';
@@ -23,7 +28,7 @@ export class SearchService {
     return response.map(transformSearchSuggestion);
   }
 
-  static async suggestPeople(query: string): Promise<SuggestedAuthor[]> {
+  static async suggestPeople(query: string): Promise<AuthorSuggestion[]> {
     const response = await ApiClient.get<any>(
       `${this.PEOPLE_SUGGEST_PATH}/?suggestion_phrases__completion=${encodeURIComponent(query)}`
     );
