+++ types/search.ts
@@ -1,6 +1,7 @@
 import { createTransformer, BaseTransformed } from './transformer';
 import { buildWorkUrl, buildAuthorUrl } from '@/utils/url';
 import { AuthorProfile } from './authorProfile';
+import { ID } from './root';
 
 export type SuggestionSource = 'api' | 'recent' | 'researchhub' | 'openalex';
 export type EntityType = 'user' | 'paper' | 'author' | 'post';
@@ -129,3 +130,52 @@ export const transformSearchSuggestion = createTransformer<any, SearchSuggestion
       };
   }
 });
+
+export interface AuthorSuggestion {
+  id: ID;
+  fullName?: string;
+  profileImage?: string;
+  institutions: Array<{
+    id: ID;
+    name: string;
+  }>;
+  headline?: string;
+  reputationHubs: string[];
+  userId?: ID;
+  createdDate?: string;
+  education: string[];
+}
+
+export const transformAuthorSuggestion = (raw: any): AuthorSuggestion => {
+  return {
+    id: raw.id,
+    fullName: raw.full_name || `${raw.first_name} ${raw.last_name}`.trim(),
+    profileImage: raw.profile_image,
+    institutions: Array.isArray(raw.institutions) ? raw.institutions : [],
+    education: Array.isArray(raw.education) ? raw.education : [],
+    headline: raw?.headline?.title,
+    reputationHubs: Array.isArray(raw.reputation_hubs) ? raw.reputation_hubs : [],
+    userId: raw.user_id,
+    createdDate: raw.created_date,
+  };
+};
+
+export const transformAuthorSuggestions = (raw: any): AuthorSuggestion[] => {
+  const authorSuggestions: AuthorSuggestion[] = [];
+  const suggestions = raw.suggestion_phrases__completion;
+
+  if (Array.isArray(suggestions)) {
+    suggestions.forEach((suggestion: any) => {
+      if (suggestion.options && Array.isArray(suggestion.options)) {
+        suggestion.options.forEach((option: any) => {
+          if (option._source) {
+            const parsed = transformAuthorSuggestion(option._source);
+            authorSuggestions.push(parsed);
+          }
+        });
+      }
+    });
+  }
+
+  return authorSuggestions;
+};
